PromptEngineering/pyproject.toml
[project]
name = "task03"
version = "0.1.0"
description = "CLI Command Generator using Grok AI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "gradio>=6.6.0",
    "httpx>=0.25.0",
    "openpyxl>=3.1.5",
]
